Added global key for FTASyncDidSync Notification. Also added an `initialSync` property
`initialSync` property is to keep track of whether or not Core Data should already have data, or if FTASync is populating the Database for the first time. It doesn't do much to set the property to `YES`, but it should set it to know after a normal sync. One could add a Core Data check elsewhere (like in AppDelegate) and set `initialSync` on the sync handler shared instance.