Boolean attributes were not being properly converted to/from NSNumbers
There is a different between `[NSNumber numberWithInteger:0]` and `[NSNumber numberWithBool:NO]`. The Later is stored in NSNumber as a CFBoolean value and the Parse framework picks up on the difference, setting the database column as either a number or boolean type. Parse framework also throws an error if you try to upload the wrong type of NSNumber for the Parse.com column type.